+++ tests_django/tests/test_metadata_store.py
@@ -1,11 +1,15 @@
 """Metadata store tests, already tested in full process so this is only unit testing."""
 from copy import deepcopy
+import datetime
 import re
 
+from django.core.cache import InvalidCacheBackendError, cache as default_cache, caches
+from django.utils import timezone
+
 import pytest
+from social_django.utils import load_backend, load_strategy
 
-from django.core.cache import cache as default_cache, InvalidCacheBackendError, caches
-
+from social_edu_federation.backends.saml_fer import FERSAMLIdentityProvider
 from social_edu_federation.django.metadata_store import CachedMetadataStore
 from social_edu_federation.parser import FederationMetadataParser
 
@@ -37,15 +41,24 @@ class MockedBackend:
         """Boilerplate to return a fixed URL"""
         return "https://domain.test/metadata/"
 
-    def setting(self, name, default_value=None):
+    def setting(self, name, default_value=None):  # pylint: disable=unused-argument
+        """
+        Defines a dummy method to return the cache name,
+        we assert this is the only setting used.
+        """
         assert name == "DJANGO_CACHE"
         return self.cache_name
 
 
 @pytest.fixture(name="cache_settings")
 def cache_settings_fixture(settings):
-    # We need to store the original value because we want to
-    # set it again before reloading the cache module at the end.
+    """
+    Pytest fixture to override cache settings.
+
+    We need to store the original value because we want to
+    set it again before reloading the initial values of the
+    cache module at the end.
+    """
     cache_initial_value = deepcopy(settings.CACHES)
 
     settings.CACHES = {
@@ -107,8 +120,11 @@ def test_fetch_remote_metadata(cache_settings, mocker):
         ),
     ],
 )
-def test_get_idp(cache_name, cache_settings, mocker):
-    """Tests `get_idp` method."""
+def test_get_idp(cache_name, cache_settings, freezer, mocker):
+    """
+    Tests `get_idp` method with different values for cache name setting (not defined and defined).
+    """
+    now = timezone.now()
     store = CachedMetadataStore(MockedBackend(cache_name=cache_name))
 
     get_metadata_mock = mocker.patch.object(FederationMetadataParser, "get_metadata")
@@ -138,19 +154,20 @@ def test_get_idp(cache_name, cache_settings, mocker):
     assert not hasattr(magic_instance, "key3")
 
     cache_to_test = default_cache if cache_name is None else caches[cache_name]
-    assert cache_to_test.get("mocked-backend:some-idp") == {
+    assert cache_to_test.get("edu_federation:mocked-backend:some-idp") == {
         "key1": "value1",
         "key2": "value2",
     }
-    assert cache_to_test.get("mocked-backend:other-idp") == {
+    assert cache_to_test.get("edu_federation:mocked-backend:other-idp") == {
         "key3": "value3",
     }
-    assert cache_to_test.get("mocked-backend:all_idps") == {
-        'other-idp': {'key3': 'value3'},
-        'some-idp': {'key1': 'value1', 'key2': 'value2'},
+    assert cache_to_test.get("edu_federation:mocked-backend:all_idps") == {
+        "other-idp": {"key3": "value3"},
+        "some-idp": {"key1": "value1", "key2": "value2"},
     }
 
     # Now call it again and assert cache is used
+    get_metadata_mock.reset_mock()
     parse_metadata_mock.reset_mock()
 
     magic_instance_again = store.get_idp("some-idp")
@@ -159,13 +176,118 @@ def test_get_idp(cache_name, cache_settings, mocker):
     assert magic_instance_again.key2 == "value2"
     assert not hasattr(magic_instance_again, "key3")
 
+    assert not get_metadata_mock.called
     assert not parse_metadata_mock.called
+
+    # Move after the cache expiration
+    freezer.move_to(now + datetime.timedelta(hours=24, minutes=1, seconds=1))
+
+    assert cache_to_test.get("edu_federation:mocked-backend:some-idp") is None
+    assert cache_to_test.get("edu_federation:mocked-backend:other-idp") is None
+    assert cache_to_test.get("edu_federation:mocked-backend:all_idps") is None
+
+    magic_instance_refreshed = store.get_idp("some-idp")
+
+    assert magic_instance_refreshed.key1 == "value1"
+    assert magic_instance_refreshed.key2 == "value2"
+    assert not hasattr(magic_instance_refreshed, "key3")
+
+    get_metadata_mock.assert_called_once_with(
+        "https://domain.test/metadata/", timeout=10
+    )
+    parse_metadata_mock.assert_called_once_with(b"been called")
 
 
 def test_get_idp_named_cache_does_not_exist(settings):
     """Tests `get_idp` method when the setting for cache is not a valid one."""
     with pytest.raises(
-            InvalidCacheBackendError,
-            match=re.escape("'invalid' does not exist in ['default']"),
+        InvalidCacheBackendError,
+        match=re.escape("'invalid' does not exist in ['default']"),
     ):
         CachedMetadataStore(MockedBackend(cache_name="invalid"))
+
+
+def test_saml_fer_backend_integration(cache_settings, mocker, settings):
+    """
+    Tests the metadata store with a real backend (`FERSAMLAuth`),
+    without cache name setting (`DJANGO_CACHE`) defined.
+    """
+    settings.AUTHENTICATION_BACKENDS = (
+        "social_edu_federation.backends.saml_fer.FERSAMLAuth",
+    )
+    settings.SOCIAL_AUTH_SAML_FER_FEDERATION_SAML_METADATA_URL = (
+        "https://domain.test/metadata/"
+    )
+
+    strategy = load_strategy()
+    backend = load_backend(strategy, "saml_fer", None)
+
+    store = CachedMetadataStore(backend)
+
+    get_metadata_mock = mocker.patch.object(FederationMetadataParser, "get_metadata")
+    get_metadata_mock.return_value = b"been called"
+    parse_metadata_mock = mocker.patch.object(
+        FederationMetadataParser, "parse_federation_metadata"
+    )
+    parse_metadata_mock.return_value = {
+        "some-idp": {
+            "name": "some-idp",
+            "entityId": "http://some-idp/",
+            "singleSignOnService": {
+                "url": "http://some-idp/ls/",
+                "binding": "urn:oasis:names:tc:SAML:2.0:bindings:HTTP-Redirect",
+            },
+            "singleLogoutService": {
+                "url": "http://some-idp/slo/",
+                "binding": "urn:oasis:names:tc:SAML:2.0:bindings:HTTP-Redirect",
+            },
+            # certificate is not validated on creation...
+            "x509cert": "MIIC4DCCAcigAwIBAgIQG...CQZXu/agfMc/tY+miyrD0=",
+            "edu_fed_data": {
+                "display_name": "Some IdP",
+            },
+        },
+        "other-idp": {
+            "name": "other-idp",
+            "entityId": "http://other-idp/",
+            "singleSignOnService": {
+                "url": "http://other-idp/ls/",
+                "binding": "urn:oasis:names:tc:SAML:2.0:bindings:HTTP-Redirect",
+            },
+            "singleLogoutService": {
+                "url": "http://other-idp/slo/",
+                "binding": "urn:oasis:names:tc:SAML:2.0:bindings:HTTP-Redirect",
+            },
+            # certificate is not validated on creation...
+            "x509cert": "MIIC4DCCAcigAwIBAgIQG...CQZXu/agfMc/tY+miyrD0=",
+            "edu_fed_data": {
+                "display_name": "Other IdP",
+            },
+        },
+    }
+
+    idp_configuration = store.get_idp("some-idp")
+    assert isinstance(idp_configuration, FERSAMLIdentityProvider)
+
+    get_metadata_mock.assert_called_once_with(
+        "https://domain.test/metadata/", timeout=10
+    )
+    parse_metadata_mock.assert_called_once_with(b"been called")
+
+    assert idp_configuration.name == "some-idp"
+    assert idp_configuration.conf == {
+        "attr_email": "urn:oid:0.9.2342.19200300.100.1.3",
+        "attr_first_name": "urn:oid:2.5.4.42",
+        "attr_full_name": "urn:oid:2.16.840.1.113730.3.1.241",
+        "attr_last_name": "urn:oid:2.5.4.4",
+        "attr_role": "urn:oid:1.3.6.1.4.1.5923.1.1.1.1",
+        "attr_user_permanent_id": "urn:oid:1.3.6.1.4.1.5923.1.1.1.6",
+        "attr_username": "urn:oid:0.9.2342.19200300.100.1.3",
+        #
+        "entity_id": "http://some-idp/",
+        "slo_url": "http://some-idp/slo/",
+        "url": "http://some-idp/ls/",
+        "x509cert": "MIIC4DCCAcigAwIBAgIQG...CQZXu/agfMc/tY+miyrD0=",
+        "x509certMulti": None,
+        "edu_fed_display_name": "Some IdP",
+    }